Prepare prefetching vv
